PRO-3340 : Fixed bug with IHT values being incorrect

+++ src/test/java/uk/gov/hmcts/probate/model/ccd/InheritanceTaxTest.java
@@ -14,7 +14,7 @@ public class InheritanceTaxTest {
     public void shouldGetNetValueInPounds() {
 
         InheritanceTax inheritanceTax = InheritanceTax.builder().formName("FORM_NAME")
-                .grossValue(1000F).netValue(800F).build();
+                .grossValue(BigDecimal.valueOf(1000F)).netValue(BigDecimal.valueOf(800F)).build();
 
         assertThat(inheritanceTax.getNetValueInPounds(), comparesEqualTo(new BigDecimal(8.00)));
     }
